requirements.txt
numpy==1.21.5
opencv-python==4.5.3.56
pyaml==21.10.1
PySide6==6.5.2
scipy==1.7.1